Adjustable Popup Modal Height
Currently, the popup modal's (`COl` class) height is fixed. If the content
is short (height), the remaining bottom of the popup modal is left with empty
space. Not only is this visually odd, but this could give users on a small
device (eg, phone) of a certain setup, give a false impression that there are
still more content, when there is none.

To avoid this, we make the popup modal's height adjustable with a maximum
height.